last few style changes - code cleanup
