+++ packages/portal-proto/src/features/annotations/AnnotationSummary.tsx
@@ -163,7 +163,7 @@ const AnnotationSummary: React.FC<AnnotationSummaryProps> = ({
         headerTitleLeft="Annotation"
         headerTitle={annotationId}
       />
-      <div className="mx-4 mt-24 mb-4">
+      <div className="mx-4 mt-6 mb-4">
         <HeaderTitle>Summary</HeaderTitle>
         <div
           data-testid="table-summary-annotation-summary"
