+++ packages/portal-proto/src/features/projects/ProjectView.tsx
@@ -1,33 +1,23 @@
-import React, { useState } from "react";
-import {
-  AnnotationDefaults,
-  GdcApiData,
-  ProjectDefaults,
-  useCoreDispatch,
-} from "@gff/core";
+import React from "react";
+import { AnnotationDefaults, GdcApiData, ProjectDefaults } from "@gff/core";
 import { FaUser, FaFile, FaEdit } from "react-icons/fa";
-import { FiDownload as DownloadIcon } from "react-icons/fi";
 import { SummaryHeader } from "@/components/Summary/SummaryHeader";
-import { Button, Loader, Tooltip } from "@mantine/core";
 import CategoryTableSummary from "@/components/Summary/CategoryTableSummary";
 import { HeaderTitle } from "@/components/tailwindComponents";
-import { DropdownWithIcon } from "@/components/DropdownWithIcon/DropdownWithIcon";
 import { HorizontalTable } from "@/components/HorizontalTable";
 import { SingularOrPluralSpan } from "@/components/SingularOrPluralSpan/SingularOrPluralSpan";
-import download from "src/utils/download";
 import PrimarySiteTable from "./PrimarySiteTable";
 import {
   formatDataForDataCategoryTable,
   formatDataForExpCategoryTable,
   formatDataForSummary,
 } from "./utils";
-import SaveCohortModal from "@/components/Modals/SaveCohortModal";
-import { focusStyles } from "@/utils/index";
 import AnnotationsTable from "./AnnotationsTable";
 import ProjectsIcon from "public/user-flow/icons/summary/projects.svg";
 import useScrollToHash from "@/hooks/useScrollToHash";
 import { useViewportSize } from "@mantine/hooks";
 import { LG_BREAKPOINT } from "../cases/utils";
+import SummaryHeaderLeft from "./SummaryHeaderLeft";
 
 export interface ProjectViewProps extends ProjectDefaults {
   readonly annotation: GdcApiData<AnnotationDefaults>;
@@ -38,20 +28,12 @@ export interface ProjectViewProps extends ProjectDefaults {
 export const ProjectView: React.FC<ProjectViewProps> = (
   projectData: ProjectViewProps,
 ) => {
-  const dispatch = useCoreDispatch();
-  const [manifestDownloadActive, setManifestDownloadActive] = useState(false);
-  const [clinicalDownloadActive, setClinicalDownloadActive] = useState(false);
-  const [biospecimenDownloadActive, setBiospecimenDownloadActive] =
-    useState(false);
-  const [showSaveCohort, setShowSaveCohort] = useState(false);
   const { width } = useViewportSize();
   useScrollToHash(["annotations"]);
 
   const Cases = (
-    <span className="flex items-center gap-0.5">
-      <div className="text-sm 2xl:text-xl">
-        <FaUser />
-      </div>
+    <span className="flex items-center gap-1">
+      <FaUser />
 
       <SingularOrPluralSpan
         customDataTestID="text-case-count-project-summary"
@@ -62,10 +44,8 @@ export const ProjectView: React.FC<ProjectViewProps> = (
   );
 
   const Files = (
-    <span className="flex items-center gap-0.5">
-      <div className="text-sm 2xl:text-xl">
-        <FaFile />
-      </div>
+    <span className="flex items-center gap-1">
+      <FaFile />
 
       <SingularOrPluralSpan
         customDataTestID="text-file-count-project-summary"
@@ -77,7 +57,7 @@ export const ProjectView: React.FC<ProjectViewProps> = (
 
   const Annotations = (
     <span className="flex items-center gap-1">
-      <FaEdit size={24} />
+      <FaEdit />
       {projectData.annotation.pagination.total > 0 ? (
         <a
           data-testid="text-annotation-count-project-summary"
@@ -116,123 +96,6 @@ export const ProjectView: React.FC<ProjectViewProps> = (
     </p>
   ) : null;
 
-  const handleBiospeciemenTSVDownload = () => {
-    setBiospecimenDownloadActive(true);
-    download({
-      endpoint: "biospecimen_tar",
-      method: "POST",
-      dispatch,
-      params: {
-        filename: `biospecimen.project-${projectData.project_id}.${new Date()
-          .toISOString()
-          .slice(0, 10)}.tar.gz`,
-        filters: {
-          op: "in",
-          content: {
-            field: "cases.project.project_id",
-            value: [projectData.project_id],
-          },
-        },
-        size: projectData.summary?.case_count,
-      },
-      done: () => setBiospecimenDownloadActive(false),
-    });
-  };
-
-  const handleBiospeciemenJSONDownload = () => {
-    setBiospecimenDownloadActive(true);
-    download({
-      endpoint: "biospecimen_tar",
-      method: "POST",
-      dispatch,
-      params: {
-        format: "JSON",
-        pretty: true,
-        filename: `biospecimen.project-${projectData.project_id}.${new Date()
-          .toISOString()
-          .slice(0, 10)}.json`,
-        filters: {
-          op: "in",
-          content: {
-            field: "cases.project.project_id",
-            value: [projectData.project_id],
-          },
-        },
-        size: projectData.summary?.case_count,
-      },
-      done: () => setBiospecimenDownloadActive(false),
-    });
-  };
-
-  const handleClinicalTSVDownload = () => {
-    setClinicalDownloadActive(true);
-    download({
-      endpoint: "clinical_tar",
-      method: "POST",
-      dispatch,
-      params: {
-        filename: `clinical.project-${projectData.project_id}.${new Date()
-          .toISOString()
-          .slice(0, 10)}.tar.gz`,
-        filters: {
-          op: "in",
-          content: {
-            field: "cases.project.project_id",
-            value: [projectData.project_id],
-          },
-        },
-        size: projectData.summary?.case_count,
-      },
-      done: () => setClinicalDownloadActive(false),
-    });
-  };
-
-  const handleClinicalJSONDownload = () => {
-    setClinicalDownloadActive(true);
-    download({
-      endpoint: "clinical_tar",
-      method: "POST",
-      dispatch,
-      params: {
-        format: "JSON",
-        pretty: true,
-        filename: `clinical.project-${projectData.project_id}.${new Date()
-          .toISOString()
-          .slice(0, 10)}.json`,
-        filters: {
-          op: "in",
-          content: {
-            field: "cases.project.project_id",
-            value: [projectData.project_id],
-          },
-        },
-        size: projectData.summary?.case_count,
-      },
-      done: () => setClinicalDownloadActive(false),
-    });
-  };
-
-  const handleManifestDownload = () => {
-    setManifestDownloadActive(true);
-    download({
-      endpoint: "files",
-      method: "POST",
-      dispatch,
-      params: {
-        filters: {
-          op: "in",
-          content: {
-            field: "cases.project.project_id",
-            value: [projectData.project_id],
-          },
-        },
-        return_type: "manifest",
-        size: 10000,
-      },
-      done: () => setManifestDownloadActive(false),
-    });
-  };
-
   const summaryData = formatDataForSummary(projectData);
   const [leftColumnData, rightColumnData] = [
     summaryData.slice(0, summaryData.length === 4 ? 2 : 3),
@@ -246,123 +109,9 @@ export const ProjectView: React.FC<ProjectViewProps> = (
         headerTitleLeft="Project"
         headerTitle={projectData.project_id}
         isModal={projectData.isModal}
-        leftElement={
-          <div className="flex gap-2">
-            <Tooltip
-              label={`Save a new cohort of ${projectData.project_id} cases`}
-              withArrow
-            >
-              <Button
-                data-testid="button-save-new-cohort-project-summary"
-                color="primary"
-                variant="outline"
-                className={`bg-base-max border-primary font-medium text-sm ${focusStyles}`}
-                onClick={() => setShowSaveCohort(true)}
-              >
-                Save New Cohort
-              </Button>
-            </Tooltip>
-
-            <SaveCohortModal
-              opened={showSaveCohort}
-              filters={{
-                mode: "and",
-                root: {
-                  "cases.project.project_id": {
-                    operator: "includes",
-                    field: "cases.project.project_id",
-                    operands: [projectData.project_id],
-                  },
-                },
-              }}
-              onClose={() => setShowSaveCohort(false)}
-            />
-
-            <DropdownWithIcon
-              customDataTestId="button-biospecimen-project-summary"
-              dropdownElements={[
-                {
-                  title: "TSV",
-                  icon: <DownloadIcon size={16} aria-label="download" />,
-                  onClick: handleBiospeciemenTSVDownload,
-                },
-                {
-                  title: "JSON",
-                  icon: <DownloadIcon size={16} aria-label="download" />,
-                  onClick: handleBiospeciemenJSONDownload,
-                },
-              ]}
-              TargetButtonChildren={
-                <span className="font-medium text-sm">
-                  {biospecimenDownloadActive ? "Processing" : "Biospecimen"}
-                </span>
-              }
-              LeftSection={
-                biospecimenDownloadActive ? (
-                  <Loader size={20} />
-                ) : (
-                  <DownloadIcon size="1rem" aria-label="download" />
-                )
-              }
-            />
-            <DropdownWithIcon
-              customDataTestId="button-clinical-project-summary"
-              dropdownElements={[
-                {
-                  title: "TSV",
-                  icon: <DownloadIcon size={16} aria-label="download" />,
-                  onClick: handleClinicalTSVDownload,
-                },
-                {
-                  title: "JSON",
-                  icon: <DownloadIcon size={16} aria-label="download" />,
-                  onClick: handleClinicalJSONDownload,
-                },
-              ]}
-              TargetButtonChildren={
-                <span className="font-medium text-sm">
-                  {clinicalDownloadActive ? "Processing" : "Clinical"}
-                </span>
-              }
-              LeftSection={
-                clinicalDownloadActive ? (
-                  <Loader size={20} />
-                ) : (
-                  <DownloadIcon size="1rem" aria-label="download" />
-                )
-              }
-            />
-            <Tooltip
-              transitionProps={{ duration: 200, transition: "fade" }}
-              w={220}
-              label="Download a manifest for use with the GDC Data Transfer Tool. The GDC
-              Data Transfer Tool is recommended for transferring large volumes of data."
-              arrowSize={10}
-              position="bottom"
-              multiline
-              withArrow
-            >
-              <Button
-                data-testid="button-manifest-project-summary"
-                variant="outline"
-                leftSection={
-                  manifestDownloadActive ? (
-                    <Loader size={20} />
-                  ) : (
-                    <DownloadIcon size="1.25em" aria-label="download" />
-                  )
-                }
-                className={`text-primary bg-base-max border-primary hover:bg-primary-darkest hover:text-base-max ${focusStyles}`}
-                classNames={{ label: "font-medium text-sm" }}
-                onClick={handleManifestDownload}
-              >
-                {manifestDownloadActive ? "Processing" : "Manifest"}
-              </Button>
-            </Tooltip>
-          </div>
-        }
+        leftElement={<SummaryHeaderLeft projectData={projectData} />}
         rightElement={
-          <div className="flex items-center gap-2 text-lg xl:text-sm 2xl:text-lg text-base-lightest leading-4 font-montserrat uppercase whitespace-no-wrap">
+          <div className="flex items-center gap-2 text-sm md:text-xl xl:text-sm 2xl:text-xl text-base-lightest leading-4 font-montserrat uppercase whitespace-no-wrap">
             Total of {Cases} {Files} {Annotations}
           </div>
         }
@@ -419,7 +168,7 @@ export const ProjectView: React.FC<ProjectViewProps> = (
         )}
         {projectData?.annotation?.pagination.count > 0 && (
           <div
-            className={`mb-16 ${
+            className={`mt-8 mb-16 ${
               projectData.isModal ? "scroll-mt-36" : "scroll-mt-72"
             }`}
             id="annotations"
